==> src/06_bert/00_get_embedding_157k.ipynb <==
# Embedding of 157k repo description

#pgadmin
import os
import psycopg2 as pg


#bert
from sentence_transformers import SentenceTransformer, util
import torch

import pandas as pd
import numpy as np

import re

import nltk
nltk.download("punkt")

from nltk import tokenize

import scipy

import datetime

#embedder = SentenceTransformer('paraphrase-MiniLM-L6-v2') #quicker model
embedder = SentenceTransformer('paraphrase-mpnet-base-v2') #most accurate, long run time

repo_data = pd.read_csv("/home/zz3hs/git/dspg21oss/data/dspg21oss/clean_github_repos_157k.csv")
len(repo_data)

# get a list of repo descriptions
repo_description = repo_data["description"].tolist()

print(repo_description[0:10])
len(repo_description)

repo_description = [x for x in repo_description if type(x) != float]
len(repo_description)

#15 cores, max memory
t1 = datetime.datetime.now()
print(t1)
repo_embedding = embedder.encode(repo_description, show_progress_bar=True, convert_to_tensor=True) 

t2 = datetime.datetime.now()
print(t2)
print(t2-t1)

repo_embedding

np_repo_embedding = np.array(repo_embedding)

np_repo_embedding

#np.save('/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_description_embeddingsXXX.npy', np_repo_embedding)

# Demo

from sentence_transformers import SentenceTransformer, util
model = SentenceTransformer('paraphrase-MiniLM-L6-v2')



#Sentences are encoded by calling model.encode()
emb1 = model.encode("Transformers: State-of-the-art Natural Language Processing for Pytorch, TensorFlow, and JAX.")
emb2 = model.encode("An  Natural Language Processing library for building bots, with entity extraction, sentiment analysis, automatic language identify, and so more.")
emb3 = model.encode("An implementation of the Grammar of Graphics in R")


cos_sim = util.cos_sim(emb1, emb2)
print("Cosine-Similarity:", cos_sim)

cos_sim = util.cos_sim(emb2, emb3)
print("Cosine-Similarity:", cos_sim)

from sentence_transformers import SentenceTransformer, util
model = SentenceTransformer('paraphrase-MiniLM-L6-v2')

sentences = ['A man is eating food.',
          'A man is eating a piece of bread.',
          'The girl is carrying a baby.',
          'A man is riding a horse.',
          'A woman is playing violin.',
          'Two men pushed carts through the woods.',
          'A man is riding a white horse on an enclosed ground.',
          'A monkey is playing drums.',
          'Someone in a gorilla costume is playing a set of drums.'
          ]

#Encode all sentences
embeddings = model.encode(sentences)

#Compute cosine similarity between all pairs
cos_sim = util.cos_sim(embeddings, embeddings)

#Add all pairs to a list with their cosine similarity score
all_sentence_combinations = []
for i in range(len(cos_sim)-1):
    for j in range(i+1, len(cos_sim)):
        all_sentence_combinations.append([cos_sim[i][j], i, j])

#Sort list by the highest cosine similarity score
all_sentence_combinations = sorted(all_sentence_combinations, key=lambda x: x[0], reverse=True)

print("Top-5 most similar pairs:")
for score, i, j in all_sentence_combinations[0:5]:
    print("{} \t {} \t {:.4f}".format(sentences[i], sentences[j], cos_sim[i][j]))
