==> src/06_bert/01_bert_init_CO.ipynb <==
# BERT on Repo Description

1. Construct a sentence corpus for each software type using labeled/manually validated repo descriptions
2. Calculate embeddingfor each corpus
3. Compare all repo description with each software type corpus using cosine-similarity score
    - It took about 1.5 hrs to run the embedding on repo data

Author: Cierra and Crystal

%reset

#pgadmin
import os
import psycopg2 as pg


#bert
from sentence_transformers import SentenceTransformer, util
import torch

import pandas as pd

import re

import nltk
nltk.download("punkt")

from nltk import tokenize

import scipy

import datetime

import math

# Embedding Model

# test cuda set up

import torch
torch.cuda.is_available()
#This should return True
torch.cuda.current_device()
# #There should be a number here
torch.cuda.device(0)
# #There should be <torch.cuda.device at 0x7efce0b03be0>
torch.cuda.device_count()
# #This should be 2 I guess

#embedder = SentenceTransformer('paraphrase-MiniLM-L6-v2') #quicker model
embedder = SentenceTransformer('paraphrase-mpnet-base-v2', device='cpu') #most accurate, long run time

# Data

## I. Unlabelled Repo Data

repo_data = pd.read_csv("~/git/dspg21oss/data/dspg21oss/clean_eng_github_repos_157k.csv")

# get a list of repo descriptions
repo_description = repo_data["description"].tolist()

## II. Labelled Repo Data -- Software Type Corpus

python_data = pd.read_excel('~/git/dspg21oss/data/dspg21oss/labelled_repo/oss_software_labelled_python_sz.xlsx') #import csv
c_data = pd.read_excel('~/git/dspg21oss/data/dspg21oss/labelled_repo/oss_software_labelled_c_sz.xlsx')
java_data = pd.read_excel('~/git/dspg21oss/data/dspg21oss/labelled_repo/oss_software_labelled_java_sz.xlsx')

# software type
type_name =  "python_label"
# filter 500 validated repos that are labelled 1 (numeric)
corpus_type_i = python_data[python_data[type_name] ==1][["slug",type_name]]

# perform a left merge to get cleaned repo description
corpus_type_i = corpus_type_i.merge(repo_data, on='slug', how='left')
# get just the description
corpus_type_i = corpus_type_i["description"].tolist()

validated_data = pd.read_csv('~/git/dspg21oss/data/dspg21oss/oss_software_labelled.csv') #import csv

# Embedding 

# get embeddings for full list of unvalidated repo descriptions

queries = repo_description

query_embeddings = []

for query in queries: 
    #Compute embeddings
    embedding = embedder.encode(query, show_progress_bar=False, convert_to_tensor=True) 
    query_embeddings.append(embedding)


types = ['ai_label', 'blockchain_label','clang_label', 'database_label',
         'dataviz_label','java_label','javascript_label','php_label','python_label']

# pre-specified number of sentences
num_sentences = 10 #find 10 most similar sentences from the corpus

# print start time
t1 = datetime.datetime.now()
print("Start:", t1)

# for each software type
for sw in types:
    # init a result list for scores
    result = []
    
    # grab the repos validated as sw type
    corpus_type = validated_data[validated_data[sw]==1][["slug",sw]]
    # left merge to get clean repo
    corpus_type = corpus_type.merge(repo_data, on='slug', how='left')
    # get just the description
    corpus_type = corpus_type["description"].tolist()
    # get rid of floats
    corpus_type = [x for x in corpus_type if type(x) == str]
    
    corpus_type_embeddings = embedder.encode(corpus_type, show_progress_bar=False)
    
    # print time for each type
    t2 = datetime.datetime.now()
    print("type ", sw,": ", t2)
    
    # for each unvalided repo query embedding 
    for query_embedding in query_embeddings:
        # use cosine-similarity and torch.topk to find the highest k scores
        cos_scores = util.pytorch_cos_sim(query_embedding, corpus_type_embeddings)[0]

        top_results = torch.topk(cos_scores, k=num_sentences)   #get the top k scores
        result.append(top_results.values.tolist()) #unlist the top result list
    
    # add results to repo_data
    result_col = sw.replace("label","sim_score")
    repo_data[result_col] = result
    
t3 =  datetime.datetime.now()
print("Finished", len(result), "descriptions at", t3)
print("It took", t3-t1, "to run.")

#save csv
repo_data.to_csv(r'/home/dab3dj/git/dspg21oss/data/dspg21oss/full_repo_sim_scores_co.csv', index = False)   


# Similarity Score Analysis

from scipy import stats
from scipy.stats import skew
import statistics #calculate mean and others

#read in data
repo_data = pd.read_csv('/home/dab3dj/git/dspg21oss/data/dspg21oss/full_repo_sim_scores_co.csv')   


repo_data.head()

# save copy of original j in case
repo_data_copy = repo_data

# list of sim score columns
score_cols = ["ai_sim_score", "blockchain_sim_score", 'clang_sim_score', 'database_sim_score', 'dataviz_sim_score',
       'java_sim_score', 'javascript_sim_score', 'php_sim_score','python_sim_score']

# change string of scores to list of scores
for col in score_cols:
    #score is in a string, convert to a list, also make sure the numbers are float
    score_ls = repo_data[col]

    score_ls_float = []
    for sentence_score in score_ls:
        sentence_score = str(sentence_score)[1:-1]
        sentence_score = sentence_score.split(",")
        item_float= []
        for item in sentence_score:
            item_float.append(float(item))
        score_ls_float.append(item_float)

    #new_col = col + "_float"
    repo_data[col] = score_ls_float

# score stats

# for each col
for col in score_cols:
    
    score_ls = repo_data[col]
    
    # get stats
    mean_score= []
    range_score = []
    max_score = []
    median_score = []
    skewness_score = []
    for sentence_score in score_ls:
        mean_score.append(statistics.mean(sentence_score))
        range_score.append(max(sentence_score)- min(sentence_score))
        max_score.append(max(sentence_score))
        median_score.append(statistics.median(sentence_score))
        skewness_score.append(stats.skew(sentence_score))
    
    # save stat columns
    col_name = col.replace("sim_score","")
    repo_data[col_name+"_mean_score"]=mean_score
    repo_data[col_name+"range_score"]=range_score
    repo_data[col_name+"max_score"]=max_score
    repo_data[col_name+"median_score"]=median_score
    repo_data[col_name+"skewness_score"]=skewness_score

repo_data.head()

import csv

# save csv 
repo_data.to_csv('/home/dab3dj/git/dspg21oss/data/dspg21oss/repo_sim_scores_stats_co.csv', index = False)

a = pd.read_csv('/home/dab3dj/git/dspg21oss/data/dspg21oss/repo_sim_scores_stats_co.csv') 
a['ai_sim_score'][0]

a['ai_sim_score'][0][0]
