==> src/06_bert/04_embedding_pca.ipynb <==
# PCA on embeddings


%reset

#pgadmin
import os

import pandas as pd
import numpy as np

from sklearn.decomposition import PCA
from sklearn.manifold import TSNE

import datetime

## Get id, which is used to match with the embedding array's index

repo_data_key = pd.read_csv("/home/zz3hs/git/dspg21oss/data/dspg21oss/clean_github_repos_157k.csv")
repo_data_key = repo_data_key[repo_data_key.description.notnull()]

repo_data_key["id"] = range(len(repo_data_key))

repo_data_key = repo_data_key[["slug", "id"]]
print(repo_data_key)

print(repo_data_key[repo_data_key.slug == "tensorflow/tensorflow"])

## Get Embeddings as array on 155,811 english repo description

embeddings = np.load('/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_description_embeddings.npy')

embeddings.shape

# PCA on bert classified repo description embeddings

We just want to get a list of repo (slugs) that are classified into a software type. We don't need to know which software type they are classified into. 

repo_label= pd.read_csv("/home/zz3hs/git/dspg21oss/data/dspg21oss/github_repos_157k_label.csv")
repo_label = repo_label[repo_label.bert == True]
repo_label = repo_label[["slug", "bert"]]

repo_label = repo_label.drop_duplicates(subset="slug") #de-duplicate by name—> unique names in the data


repo_label[repo_label.slug == "tensorflow/tensorflow"]

repo_label

#repo_label.to_csv("/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_classified_slugs.csv", index = False)



# get the array id
repo_label_id = pd.merge(repo_label, repo_data_key, on = "slug", how = "left")

print(len(repo_label_id))
repo_label_id[repo_label_id.slug == "tensorflow/tensorflow"]


repo_ids = repo_label_id["id"].tolist()
len(repo_ids)

repo_label_embeddings = []
for repo_id in repo_ids:
    embedding = embeddings[repo_id]
    repo_label_embeddings.append(embedding)
len(repo_label_embeddings)   

### check variance explained for top 20 pc's

pca = PCA(n_components=10)
pca.fit(repo_label_embeddings)
#percent of variance explained in each pc
pca.explained_variance_

pca = PCA(n_components=3)
pca_embedding = pca.fit_transform(repo_label_embeddings)

pca_embedding.shape

embedding_df = pd.DataFrame({'x': pca_embedding[:, 0], 'y': pca_embedding[:, 1], 'z' : pca_embedding[:,2]})

embedding_df

repo_label_id["x"] = embedding_df.x
repo_label_id["y"] = embedding_df.y
repo_label_id["z"] = embedding_df.z

repo_label_id

repo_data_out = repo_label_id[repo_label_id.x.notnull()]
repo_data_out = repo_data_out[["slug", "x", "y", "z"]]
repo_data_out

#repo_data_out.to_csv("/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_classified_embedding_3d_pca.csv", index = False)

# PCA on blockchain/dataviz repo description embeddings

repo_blockchain_dataviz = pd.read_csv("/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_blockchain_dataviz.csv")
repo_blockchain_dataviz

repo_blockchain_dataviz_id = pd.merge(repo_blockchain_dataviz, repo_data, on = "slug")
repo_ids = repo_blockchain_dataviz_id["id"]
repo_ids

repo_blockchain_dataviz_id = repo_blockchain_dataviz_id.set_index('id')
repo_blockchain_dataviz_id

print(repo_data.iloc[151])

repo_blockchain_dataviz_embeddings = []
for repo_id in repo_ids:
    embedding = embeddings[repo_id]
    repo_blockchain_dataviz_embeddings.append(embedding)
    


pca = PCA(n_components=3)
pca_embedding = pca.fit_transform(repo_blockchain_dataviz_embeddings)

embedding_df = pd.DataFrame({'x': pca_embedding[:, 0], 'y': pca_embedding[:, 1], 'z' : pca_embedding[:,2]})

repo_blockchain_dataviz_id["x"] = embedding_df.x
repo_blockchain_dataviz_id["y"] = embedding_df.y
repo_blockchain_dataviz_id["z"] = embedding_df.z

repo_blockchain_dataviz_id

#repo_blockchain_dataviz_id.to_csv("/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_blockchain_dataviz_embedding_3d_pca.csv", index = False)
