==> src/06_bert/04_embedding_tsne.ipynb <==
# PCA on embeddings


%reset

#pgadmin
import os

import pandas as pd
import numpy as np

from sklearn.decomposition import PCA
from sklearn.manifold import TSNE

import datetime

repo_data = pd.read_csv("/home/zz3hs/git/dspg21oss/data/dspg21oss/clean_github_repos_157k.csv")
repo_data = repo_data[repo_data.description.notnull()]
repo_data

data = np.load('repo_description_embeddings.npy')

data.shape

# PCA

pca = PCA(n_components=2)
pca.fit(data)

#percent of variance explained in each pc
pca.explained_variance_

print(pca)

# TSNE

tsne_model = TSNE(n_components=3,perplexity = 30,n_iter=3500, init='pca', random_state=1)

t1 = datetime.datetime.now()
print("Start:", t1)

tsne_embedding = tsne_model.fit_transform(data)

t2 =  datetime.datetime.now()
print("Finished", len(result), "descriptions at", t2)
print("It took", t2-t1, "to run.")

type(tsne_embedding)

tsne_embedding.shape

embedding_df = pd.DataFrame({'x': tsne_embedding[:, 0], 'y': tsne_embedding[:, 1], 'z' : tsne_embedding[:,2]})

embedding_df

repo_data["x"] = embedding_df.x
repo_data["y"] = embedding_df.y
repo_data["z"] = embedding_df.z

repo_data.to_csv("/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_embedding_3d.csv", index = False)
